+++ PrimaryDataTypes/NosOfBitsProblem-5.1/brute-force-bit-method.h
@@ -1,10 +1,10 @@
 #ifndef BITFORCEBITMETHOD_H
 #define BITFORCEBITMETHOD_H
 
-int normalMethod(int number );
-int bitMaskMethod(int number);
-int rightShiftMethod(int number);
-int lowestBitMethod(int number);
+int normalMethod(long long number );
+int bitMaskMethod(long long number);
+int rightShiftMethod(long long number);
+int lowestBitMethod(long long number);
 
 
 #endif // BITFORCEBITMETHOD_H
